pyproject.toml
[project]
name = "isoart"
version = "0.1.0"
description = "Isometric pixel-art asset library for Python"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Henrik H", email = "rhedak00@yahoo.de" }
]
keywords = ["isometric", "pixel-art", "game-art", "pillow", "image"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Games/Entertainment",
    "Typing :: Typed",
]
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.4",
    "pillow>=12.2.0",
]

[project.urls]
Homepage = "https://github.com/rhedak/isoart"
Repository = "https://github.com/rhedak/isoart"
Issues = "https://github.com/rhedak/isoart/issues"

[build-system]
requires = ["uv_build>=0.11.3,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.15.11",
]
